Battle screen
Made the new battle, which I will begin to develop. I made a script to change the scene to the battle screen when you press the play button. I also tested out another script that when you press  the escape key, it will send you back to the main menu.