Battle screen
Drew and added all of the components of the battle screen and added it to Unity. Added an edge scrolling component and added barriers to the edge of the screen. I will figure out how to make sure my mouse doesn't go to my other screen when I want to pan over.