Added another animation to the brawler
I added the punch animation to the brawler. I'll have to figure out how to make the brawler keep walking until it detects an enemy and starts punching it. I wonder how I can classify my characters to troops and enemies. I also wonder if I can change my brawler's hitbox for the punch animation so that his arm hitbox extends only when he does a full punch. Should I make it so that it punches a single target or just the area in front of him?