Health bar mechanics
Health bar system works! The enemy (red circle) has a health bar which only shows when it is damaged. The brawler will punch and the precise frame where it does the blow will cause the enemy to lose health. The brawler should immediately stop its animation when it detects no enemies in its range UNLESS it is past the punch blow frame. Next, I'll need to work on the color gradient of the health bar. Then, I'll start working on my first enemy.